+++ implementation/js/site/context/demo.js
@@ -1,8 +1,12 @@
-Application.context('Demo', {
-    //debug: true,
-    className: 'wrapper container',
-    template: [
-        '<div region="center"></div>',
-    ],
-    navRegion: 'center'
-});
+(function(app){
+
+	app.context('Demo', {
+	    //debug: true,
+	    className: 'wrapper container',
+	    template: [
+	        '<div region="center"></div>',
+	    ],
+	    navRegion: 'center'
+	});
+
+})(Application);